+++ test/cases/specific_schema_test_sqlserver.rb
@@ -5,8 +5,14 @@ class SpecificSchemaTestSQLServer < ActiveRecord::TestCase
   after { SSTestEdgeSchema.delete_all }
 
   it 'handle dollar symbols' do
-    SSTestDollarTableName.new.save
+    SSTestDollarTableName.create!
     SSTestDollarTableName.limit(20).offset(1)
+  end
+
+  it 'handle dot table names' do
+    SSTestDotTableName.create! name: 'test'
+    SSTestDotTableName.limit(20).offset(1)
+    SSTestDotTableName.where(name: 'test').first.must_be :present?
   end
 
   it 'models can use tinyint pk tables' do