+++ test/models/sqlserver/dot_table_name.rb
@@ -1,0 +1,3 @@
+class SSTestDotTableName < ActiveRecord::Base
+  self.table_name = '[dbo].[some.Name]'
+end